+++ compiler-core/src/exhaustiveness.rs
@@ -65,8 +65,8 @@ pub mod printer;
 
 use crate::{
     ast::{
-        self, AssignName, BitArraySize, Endianness, IntOperator, TypedBitArraySize, TypedClause,
-        TypedPattern, TypedPatternBitArraySegment,
+        self, AssignName, BitArraySize, Endianness, IntOperator, SrcSpan, TypedBitArraySize,
+        TypedClause, TypedPattern, TypedPatternBitArraySegment,
     },
     strings::{
         convert_string_escape_chars, length_utf16, length_utf32, string_to_utf16_bytes,
@@ -410,7 +410,7 @@ impl Body {
     fn assign_segment_constant_value(&mut self, name: EcoString, value: &BitArrayMatchedValue) {
         let value = match value {
             BitArrayMatchedValue::LiteralFloat(value) => BoundValue::LiteralFloat(value.clone()),
-            BitArrayMatchedValue::LiteralInt(value) => BoundValue::LiteralInt(value.clone()),
+            BitArrayMatchedValue::LiteralInt { value, .. } => BoundValue::LiteralInt(value.clone()),
             BitArrayMatchedValue::LiteralString { value, .. } => {
                 BoundValue::LiteralString(value.clone())
             }
@@ -525,6 +525,33 @@ impl Pattern {
                 },
             ) if index != variant => true,
             _ => false,
+        }
+    }
+
+    fn is_matching_on_impossible_segment(&self) -> Option<Vec<ImpossibleBitArraySegmentPattern>> {
+        match self {
+            Self::BitArray { tests } => {
+                let impossible_segments = tests
+                    .iter()
+                    .filter_map(|test| match test {
+                        BitArrayTest::Size(_)
+                        | BitArrayTest::CatchAllIsBytes { .. }
+                        | BitArrayTest::ReadSizeIsNotNegative { .. }
+                        | BitArrayTest::SegmentIsFiniteFloat { .. } => None,
+
+                        BitArrayTest::Match(MatchTest { value, read_action }) => {
+                            value.is_impossible_segment(read_action)
+                        }
+                    })
+                    .collect_vec();
+
+                if impossible_segments.is_empty() {
+                    None
+                } else {
+                    Some(impossible_segments)
+                }
+            }
+            _ => None,
         }
     }
 }
@@ -1162,7 +1189,13 @@ impl MatchTest {
 #[derive(Clone, Eq, PartialEq, Debug, serde::Serialize, serde::Deserialize)]
 pub enum BitArrayMatchedValue {
     LiteralFloat(EcoString),
-    LiteralInt(BigInt),
+    LiteralInt {
+        value: BigInt,
+        /// This is carried around for better error reporting in case a segment
+        /// is deemed unreachable: it is the location this literal value comes
+        /// from in the whole pattern.
+        location: SrcSpan,
+    },
     LiteralString {
         value: EcoString,
         encoding: StringEncoding,
@@ -1182,7 +1215,7 @@ impl BitArrayMatchedValue {
     pub(crate) fn is_literal(&self) -> bool {
         match self {
             BitArrayMatchedValue::LiteralFloat(_)
-            | BitArrayMatchedValue::LiteralInt(_)
+            | BitArrayMatchedValue::LiteralInt { .. }
             | BitArrayMatchedValue::LiteralString { .. } => true,
             BitArrayMatchedValue::Variable(..) | BitArrayMatchedValue::Discard(..) => false,
             BitArrayMatchedValue::Assign { value, .. } => value.is_literal(),
@@ -1201,10 +1234,93 @@ impl BitArrayMatchedValue {
             // TODO: We could also implement the interfering optimisation for
             // literal ints as well, but that will be a bit trickier than
             // strings.
-            BitArrayMatchedValue::LiteralInt(_)
+            BitArrayMatchedValue::LiteralInt { .. }
             | BitArrayMatchedValue::LiteralFloat(_)
             | BitArrayMatchedValue::Variable(_)
             | BitArrayMatchedValue::Discard(_) => None,
+        }
+    }
+
+    /// If we can statically tell the segment will never match, this will return
+    /// the reason why it can't match. Otherwise it returns `None`.
+    ///
+    fn is_impossible_segment(
+        &self,
+        read_action: &ReadAction,
+    ) -> Option<ImpossibleBitArraySegmentPattern> {
+        match self {
+            BitArrayMatchedValue::Assign { value, .. } => value.is_impossible_segment(read_action),
+            BitArrayMatchedValue::LiteralInt { value, location } => {
+                let size = read_action.size.constant_bits()?.to_u32()?;
+                if representable_with_bits(value.clone(), size, read_action.signed) {
+                    None
+                } else {
+                    Some(ImpossibleBitArraySegmentPattern::UnrepresentableInteger {
+                        value: value.clone(),
+                        size,
+                        location: *location,
+                        signed: read_action.signed,
+                    })
+                }
+            }
+
+            BitArrayMatchedValue::LiteralFloat(_)
+            | BitArrayMatchedValue::LiteralString { .. }
+            | BitArrayMatchedValue::Variable(_)
+            | BitArrayMatchedValue::Discard(_) => None,
+        }
+    }
+
+    /// Returns true if the two `MatchedValue`s are matching for the exact same
+    /// thing. Note how this doesn't automatically mean they will be matching
+    /// for the exact same thing in a whole bit array pattern as they could be
+    /// matching at different positions!
+    ///
+    /// ```gleam
+    /// <<1, _>>
+    /// <<_, 1>>
+    /// // Both are matching on the same value but at different positions!
+    /// ```
+    ///
+    fn checking_same_value(&self, other: &Self) -> bool {
+        match (self, other) {
+            // Ints need special care: they're carrying around information about
+            // where they come from. But as for equality is concerned for the
+            // pattern match compilation, they are considered equal as long as
+            // the two values are equal, no matter where the values come from.
+            (Self::LiteralInt { value: one, .. }, Self::LiteralInt { value: other, .. }) => {
+                one == other
+            }
+            (Self::LiteralInt { .. }, _) => false,
+
+            // All the other cases follow the standard equality implementation.
+            (Self::LiteralFloat(one), Self::LiteralFloat(other)) => one == other,
+            (Self::LiteralFloat(_), _) => false,
+
+            // Two literal string matches are the same if they match for the
+            // same bytes. No matter the original starting string and encoding.
+            (Self::LiteralString { bytes: one, .. }, Self::LiteralString { bytes: other, .. }) => {
+                one == other
+            }
+            (Self::LiteralString { .. }, _) => false,
+
+            (Self::Variable(one), Self::Variable(other)) => one == other,
+            (Self::Variable(_), _) => false,
+
+            (Self::Discard(_), Self::Discard(_)) => true,
+            (Self::Discard(_), _) => false,
+
+            (
+                Self::Assign {
+                    name: _,
+                    value: one,
+                },
+                Self::Assign {
+                    name: _,
+                    value: other,
+                },
+            ) => one.checking_same_value(other),
+            (Self::Assign { .. }, _) => false,
         }
     }
 }
@@ -1221,7 +1337,7 @@ impl BitArrayMatchedValue {
         match self {
             BitArrayMatchedValue::Discard(_) => true,
             BitArrayMatchedValue::LiteralFloat(_)
-            | BitArrayMatchedValue::LiteralInt(_)
+            | BitArrayMatchedValue::LiteralInt { .. }
             | BitArrayMatchedValue::LiteralString { .. }
             | BitArrayMatchedValue::Variable(_)
             | BitArrayMatchedValue::Assign { .. } => false,
@@ -1230,6 +1346,43 @@ impl BitArrayMatchedValue {
 }
 
 impl BitArrayTest {
+    /// Wether two bit array tests are equivalent, that is they are checking
+    /// for the same thing.
+    ///
+    fn equivalent_to(&self, other: &Self) -> bool {
+        match (self, other) {
+            (BitArrayTest::Size(one), BitArrayTest::Size(other)) => one == other,
+            (BitArrayTest::Size(_), _) => false,
+
+            (
+                BitArrayTest::Match(MatchTest { value, read_action }),
+                BitArrayTest::Match(MatchTest {
+                    value: other_value,
+                    read_action: other_read_action,
+                }),
+            ) => value.checking_same_value(other_value) && read_action == other_read_action,
+            (BitArrayTest::Match(_), _) => false,
+
+            (
+                BitArrayTest::CatchAllIsBytes { size_so_far: one },
+                BitArrayTest::CatchAllIsBytes { size_so_far: other },
+            ) => one == other,
+            (BitArrayTest::CatchAllIsBytes { .. }, ..) => false,
+
+            (
+                BitArrayTest::ReadSizeIsNotNegative { size: one },
+                BitArrayTest::ReadSizeIsNotNegative { size: other },
+            ) => one == other,
+            (BitArrayTest::ReadSizeIsNotNegative { .. }, _) => false,
+
+            (
+                BitArrayTest::SegmentIsFiniteFloat { read_action: one },
+                BitArrayTest::SegmentIsFiniteFloat { read_action: other },
+            ) => one == other,
+            (BitArrayTest::SegmentIsFiniteFloat { .. }, _) => false,
+        }
+    }
+
     /// Tells us if this test is guaranteed to succeed given another test that
     /// we know has already succeeded.
     ///
@@ -1239,7 +1392,7 @@ impl BitArrayTest {
     #[must_use]
     fn succeeds_if_succeeding(&self, succeeding: &BitArrayTest) -> Confidence {
         match (succeeding, self) {
-            (one, other) if one == other => Confidence::Certain,
+            (one, other) if one.equivalent_to(other) => Confidence::Certain,
             (BitArrayTest::Size(succeeding), BitArrayTest::Size(test)) => {
                 test.succeeds_if_succeeding(succeeding)
             }
@@ -1272,7 +1425,7 @@ impl BitArrayTest {
     #[must_use]
     fn fails_if_failing(&self, failing: &BitArrayTest) -> Confidence {
         match (failing, self) {
-            (one, other) if one == other => Confidence::Certain,
+            (one, other) if one.equivalent_to(other) => Confidence::Certain,
             (BitArrayTest::Size(failing), BitArrayTest::Size(test)) => {
                 test.fails_if_failing(failing)
             }
@@ -1904,6 +2057,14 @@ impl CompileCaseResult {
             .contains(&(clause, pattern_index))
         {
             Reachability::Unreachable(UnreachablePatternReason::ImpossibleVariant)
+        } else if let Some(segments) = self
+            .diagnostics
+            .match_impossible_segments
+            .get(&(clause, pattern_index))
+        {
+            Reachability::Unreachable(UnreachablePatternReason::ImpossibleSegments(
+                segments.clone(),
+            ))
         } else {
             Reachability::Unreachable(UnreachablePatternReason::DuplicatePattern)
         }
@@ -1946,7 +2107,7 @@ impl CompiledCase {
 
 /// Whether a pattern is reachable, or why it is unreachable.
 ///
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Reachability {
     Reachable,
     Unreachable(UnreachablePatternReason),
@@ -1983,6 +2144,29 @@ pub struct Diagnostics {
     /// See `reachable` for an explanation of its structure.
     ///
     pub match_impossible_variants: HashSet<(usize, usize)>,
+
+    /// Patterns that are matching on bit array segments the compiler can tell
+    /// for sure will never match.
+    ///
+    pub match_impossible_segments: HashMap<(usize, usize), Vec<ImpossibleBitArraySegmentPattern>>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
+pub enum ImpossibleBitArraySegmentPattern {
+    UnrepresentableInteger {
+        value: BigInt,
+        size: u32,
+        location: SrcSpan,
+        signed: bool,
+    },
+}
+
+impl ImpossibleBitArraySegmentPattern {
+    pub fn location(&self) -> SrcSpan {
+        match self {
+            ImpossibleBitArraySegmentPattern::UnrepresentableInteger { location, .. } => *location,
+        }
+    }
 }
 
 impl<'a> Compiler<'a> {
@@ -1995,6 +2179,7 @@ impl<'a> Compiler<'a> {
                 missing: false,
                 reachable: HashSet::new(),
                 match_impossible_variants: HashSet::new(),
+                match_impossible_segments: HashMap::new(),
             },
         }
     }
@@ -2030,6 +2215,21 @@ impl<'a> Compiler<'a> {
             .diagnostics
             .match_impossible_variants
             .insert((branch.clause_index, branch.alternative_index));
+    }
+
+    fn mark_as_matching_impossible_segment(
+        &mut self,
+        branch: &Branch,
+        segments: Vec<ImpossibleBitArraySegmentPattern>,
+    ) {
+        let _ = self
+            .diagnostics
+            .reachable
+            .remove(&(branch.clause_index, branch.alternative_index));
+        let _ = self
+            .diagnostics
+            .match_impossible_segments
+            .insert((branch.clause_index, branch.alternative_index), segments);
     }
 
     fn compile(&mut self, mut branches: VecDeque<Branch>) -> Decision {
@@ -2173,8 +2373,14 @@ impl<'a> Compiler<'a> {
             };
 
             let checked_pattern = self.pattern(pattern_check.pattern);
+
             if checked_pattern.is_matching_on_unreachable_variant(branch_mode) {
                 self.mark_as_matching_impossible_variant(&branch);
+                continue;
+            } else if let Some(unreachable_segments) =
+                checked_pattern.is_matching_on_impossible_segment()
+            {
+                self.mark_as_matching_impossible_segment(&branch, unreachable_segments);
                 continue;
             }
 
@@ -3231,7 +3437,7 @@ impl CaseToCompile {
             // following read actions as a valid size.
             match &value {
                 BitArrayMatchedValue::LiteralFloat(_)
-                | BitArrayMatchedValue::LiteralInt(_)
+                | BitArrayMatchedValue::LiteralInt { .. }
                 | BitArrayMatchedValue::LiteralString { .. }
                 | BitArrayMatchedValue::Discard(_) => {}
                 BitArrayMatchedValue::Variable(name)
@@ -3267,7 +3473,14 @@ fn segment_matched_value(
 ) -> BitArrayMatchedValue {
     let pattern = pattern.unwrap_or(&segment.value);
     match pattern {
-        ast::Pattern::Int { int_value, .. } => BitArrayMatchedValue::LiteralInt(int_value.clone()),
+        ast::Pattern::Int {
+            int_value,
+            location,
+            ..
+        } => BitArrayMatchedValue::LiteralInt {
+            value: int_value.clone(),
+            location: *location,
+        },
         ast::Pattern::Float { value, .. } => BitArrayMatchedValue::LiteralFloat(value.clone()),
         ast::Pattern::String { value, .. } if segment.has_utf16_option() => {
             BitArrayMatchedValue::LiteralString {
@@ -3418,4 +3631,58 @@ fn superset(
             Some(occurrences) => occurrences >= other_occurrences,
             None => false,
         })
+}
+
+#[must_use]
+fn representable_with_bits(value: BigInt, bits: u32, signed: bool) -> bool {
+    // No number is representable in 0 bits.
+    if bits == 0 {
+        return false;
+    };
+    if signed {
+        // Signed numbers range in [-2^(bits-1), 2^(bits-1)[
+        let power = BigInt::from(2).pow(bits - 1);
+        -&power <= value && value < power
+    } else {
+        // Unsigned numbers range in [0, 2^bits[
+        BigInt::from(0) <= value && value < BigInt::from(2).pow(bits)
+    }
+}
+
+#[cfg(test)]
+mod representable_with_bits_test {
+    use crate::exhaustiveness::representable_with_bits;
+    use num_bigint::BigInt;
+
+    #[test]
+    fn positive_number_representable_with_bits_test() {
+        // 9 can be represented as a >=4 bits unsigned number.
+        assert!(representable_with_bits(9.into(), 4, false));
+        assert!(representable_with_bits(9.into(), 5, false));
+        // But not in <=3 bits as an unsigned number.
+        assert!(!representable_with_bits(9.into(), 3, false));
+        assert!(!representable_with_bits(9.into(), 2, false));
+
+        // It can be represented as a >=5 bit signed number.
+        assert!(representable_with_bits(9.into(), 5, true));
+        assert!(representable_with_bits(9.into(), 6, true));
+        // But not in <= 4 bits as a signed number.
+        assert!(!representable_with_bits(9.into(), 4, true));
+        assert!(!representable_with_bits(9.into(), 3, true));
+    }
+
+    #[test]
+    fn negative_number_representable_with_bits_test() {
+        // A negative number will never be representable as an unsigned number,
+        // no matter the number of bits!
+        assert!(!representable_with_bits(BigInt::from(-9), 1, false));
+        assert!(!representable_with_bits(BigInt::from(-9), 500, false));
+
+        // -9 can be represented in >=5 bits as a signed number.
+        assert!(representable_with_bits(BigInt::from(-9), 5, true));
+        assert!(representable_with_bits(BigInt::from(-9), 6, true));
+        // But not in <= 4 bits as a signed number.
+        assert!(!representable_with_bits(BigInt::from(-9), 4, true));
+        assert!(!representable_with_bits(BigInt::from(-9), 3, true));
+    }
 }